Module Analysis_and_optimization.Debug_data_generation

val json_to_mir : (Middle.Expr.Typed.t Middle.SizedType.t * 'a * string) list -> Yojson.Basic.t -> (string, Middle.Expr.Typed.t) Core.Map.Poly.t

Translates Yojson object into a data type that `gen_values_json` understands.

val gen_values_json : ?new_only:bool -> ?context:(string, Middle.Expr.Typed.t) Core.Map.Poly.t -> (Middle.Expr.Typed.t Middle.SizedType.t * Middle.Expr.Typed.t Middle.Transformation.t * string) list -> (string, Frontend.Errors.t) Core.result

Generates values matching the given declarations and formats them as a JSON string. The declarations may depend on additional values supplied in `context`. If `new_only` is true (defaults to false) the output does not include the values in `context`