Module type Monotone_framework_sigs.FLOWGRAPH

The API for a flowgraph, needed for the mfp algorithm in the monotone framework. Assumed invariants: successors contains all graph nodes as keys initials is a subset of the graph nodes

type labels

val initials : labels Std.Set.Poly.t

val successors : labels Std.Set.Poly.t Dataflow_types.LabelMap.t