Module type Monotone_framework_sigs.MONOTONE_FRAMEWORK

The API for a monotone framework. mfp computes the minimal fixed point of the equations/inequalities defined between property lattice elements at the entry and exit of different flowgraph nodes, where these equations/inequalities are generated from the transfer function. Returns a map of the (input_properties, output_properties) for each node l in the flow graph. The analysis performed is always a forward analysis. For a reverse analysis, supply the reverse flow graph.

type labels

type properties

val mfp : unit -> (labels, properties entry_exit) Core.Map.Poly.t