Module type Monotone_framework_sigs.TRANSFER_FUNCTION

The API for a transfer function, needed for the mfp algorithm in the monotone framework. This describes how output properties are computed from input properties at a given node in the flow graph.

type labels

type properties

val transfer_function : labels -> properties -> properties