Module Analysis_and_optimization.Pedantic_analysis

type warning_span = Middle.Location_span.t * string

val warn_pedantic : Middle.Program.Typed.t -> warning_span list

Collect all pedantic mode warnings and print to a string.

val warn_uninitialized : Middle.Program.Typed.t -> warning_span list

Collect warnings about each variable which is used before being initialized and print to a string.