Module Analysis_and_optimization__Dependence_analysis

type node_dep_info = {

predecessors : Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t;

parents : Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t;

reaching_defn_entry : Analysis_and_optimization.Dataflow_types.reaching_defn Core_kernel.Set.Poly.t;

reaching_defn_exit : Analysis_and_optimization.Dataflow_types.reaching_defn Core_kernel.Set.Poly.t;

meta : Middle.Location_span.t;

}

Sufficient information about each node to build the dependency graph.

Label dependence doesn't need the exit RD set, but variable dependence does.

val node_immediate_dependencies : (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t -> ?⁠blockers:Analysis_and_optimization.Dataflow_types.vexpr Core_kernel.Set.Poly.t -> Analysis_and_optimization.Dataflow_types.label -> Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t

Given dependency information for each node, find the 'immediate' dependencies of a node, where 'immediate' means the first-degree control flow parents and the reachable definitions of RHS variables.

val node_dependencies : (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t -> Analysis_and_optimization.Dataflow_types.label -> Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t

Given dependency information for each node, find all of the dependencies of a single node.

val node_vars_dependencies : (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t -> ?⁠blockers:Analysis_and_optimization.Dataflow_types.vexpr Core_kernel.Set.Poly.t -> Analysis_and_optimization.Dataflow_types.vexpr Core_kernel.Set.Poly.t -> Analysis_and_optimization.Dataflow_types.label -> Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t

Given dependency information for each node, find all of the dependencies of a set of variables at single node.

'blockers' are variables which will not be traversed.

val build_dep_info_map : Middle.Program.Typed.t -> Middle.Stmt.Located.t -> (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t

Build the dependency information for each node in the log_prob section of a program

val log_prob_build_dep_info_map : Middle.Program.Typed.t -> (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t

Build the dependency information for each node in the log_prob section of a program

val all_node_dependencies : (Analysis_and_optimization.Dataflow_types.label, (Middle.Expr.Typed.t, Analysis_and_optimization.Dataflow_types.label) Middle.Stmt.Fixed.Pattern.t * node_dep_info) Core_kernel.Map.Poly.t -> (Analysis_and_optimization.Dataflow_types.label, Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t) Core_kernel.Map.Poly.t

Given dependency information for each node, find all of the dependencies of all nodes, effectively building the dependency graph.

This is more efficient than calling node_dependencies on each node individually.

val log_prob_dependency_graph : Middle.Program.Typed.t -> (Analysis_and_optimization.Dataflow_types.label, Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t) Core_kernel.Map.Poly.t

Build the dependency graph for the log_prob section of a program, where labels correspond to the labels built by statement_map.

val reaching_defn_lookup : Analysis_and_optimization.Dataflow_types.reaching_defn Core_kernel.Set.Poly.t -> Analysis_and_optimization.Dataflow_types.vexpr -> Analysis_and_optimization.Dataflow_types.label Core_kernel.Set.Poly.t

val mir_uninitialized_variables : Middle.Program.Typed.t -> (Middle.Location_span.t * string) Core_kernel.Set.Poly.t

Produce a list of uninitialized variables and their label locations, from the flowgraph starting at the given statement