Module Analysis_and_optimization__Mir_utils

val var_declarations : ('a, 'b) Middle.Stmt.Fixed.t -> string Core_kernel.Set.Poly.t

val num_expr_value : Middle.Expr.Typed.t -> (float * string) option

type bound_values = {

lower : [ `None | `Nonlit | `Lit of float ];

upper : [ `None | `Nonlit | `Lit of float ];

}

val trans_bounds_values : Middle.Expr.Typed.t Middle.Transformation.t -> bound_values

val chop_dist_name : string -> string Core_kernel.Option.t

val top_var_declarations : Middle.Stmt.Located.t -> string Core_kernel.Set.Poly.t

val data_set : ?⁠exclude_transformed:bool -> ?⁠exclude_ints:bool -> Middle.Program.Typed.t -> string Core_kernel.Set.Poly.t

val parameter_set : ?⁠include_transformed:bool -> Middle.Program.Typed.t -> (string * Middle.Expr.Typed.t Middle.Transformation.t) Core_kernel.Set.Poly.t

val parameter_names_set : ?⁠include_transformed:bool -> Middle.Program.Typed.t -> string Core_kernel.Set.Poly.t

val fold_expr : take_expr:('c -> Middle.Expr.Typed.t -> 'c) -> init:'c -> Middle.Expr.Typed.t -> 'c

val fold_stmts : take_expr:('c -> Middle.Expr.Typed.t -> 'c) -> take_stmt:('c -> Middle.Stmt.Located.t -> 'c) -> init:'c -> Middle.Stmt.Located.t Core_kernel.List.t -> 'c

val map_rec_expr : (Middle.Expr.Typed.t Middle.Expr.Fixed.Pattern.t -> Middle.Expr.Typed.t Middle.Expr.Fixed.Pattern.t) -> Middle.Expr.Typed.t -> Middle.Expr.Typed.t

val map_rec_expr_state : ('s -> Middle.Expr.Typed.t Middle.Expr.Fixed.Pattern.t -> Middle.Expr.Typed.t Middle.Expr.Fixed.Pattern.t * 's) -> 's -> Middle.Expr.Typed.t -> Middle.Expr.Typed.t * 's

val map_rec_stmt_loc : ((Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t) -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t

val top_down_map_rec_stmt_loc : ((Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t) -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t

val map_rec_state_stmt_loc : ('s -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t * 's) -> 's -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t * 's

val map_rec_stmt_loc_num : (int, Middle.Stmt.Located.Non_recursive.t) Core_kernel.Map.Poly.t -> (int -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t) -> Middle.Stmt.Located.Non_recursive.t -> Middle.Stmt.Located.t

val map_rec_state_stmt_loc_num : (int, Middle.Stmt.Located.Non_recursive.t) Core_kernel.Map.Poly.t -> (int -> 's -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t * 's) -> 's -> Middle.Stmt.Located.Non_recursive.t -> Middle.Stmt.Located.t * 's

val stmt_loc_of_stmt_loc_num : (int, Middle.Stmt.Located.Non_recursive.t) Core_kernel.Map.Poly.t -> Middle.Stmt.Located.Non_recursive.t -> Middle.Stmt.Located.t

val statement_stmt_loc_of_statement_stmt_loc_num : (int, Middle.Stmt.Located.Non_recursive.t) Core_kernel.Map.Poly.t -> (Middle.Expr.Typed.t, int) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, Middle.Stmt.Located.t) Middle.Stmt.Fixed.Pattern.t

val unnumbered_prog_of_numbered_prog : (int, Middle.Stmt.Located.Non_recursive.t) Core_kernel.Map.Poly.t -> ('a -> 'b) -> (Middle.Stmt.Located.Non_recursive.t, 'a) Middle.Program.t -> (Middle.Stmt.Located.t, 'b) Middle.Program.t

val fwd_traverse_statement : ('e, 'a) Middle.Stmt.Fixed.Pattern.t -> init:'f -> f:('f -> 'a -> 'f * 'c) -> 'f * ('e, 'c) Middle.Stmt.Fixed.Pattern.t

A traversal that simultaneously accumulates a state (type 'f) and replaces the substatement values from ('a to 'c). Traversal is done in-order but ignores branching, e.g., and if's then block is followed by the else block rather than branching.

val vexpr_of_expr_exn : Middle.Expr.Typed.t -> Analysis_and_optimization.Dataflow_types.vexpr

Take a LHS expression from a general expression, throwing an exception if it can't be a LHS expression.

val expr_var_set : Middle.Expr.Typed.t -> (Analysis_and_optimization.Dataflow_types.vexpr * Middle.Expr.Typed.Meta.t) Core_kernel.Set.Poly.t

The set of variables in an expression, including inside an index. For use in RHS sets, not LHS assignment sets, except in a target term.

val index_var_set : Middle.Expr.Typed.t Middle.Index.t -> (Analysis_and_optimization.Dataflow_types.vexpr * Middle.Expr.Typed.Meta.t) Core_kernel.Set.Poly.t

The set of variables in an index. For use in RHS sets, not LHS assignment sets, except in a target term

val expr_var_names_set : Middle.Expr.Typed.t -> string Core_kernel.Set.Poly.t

Return the names of the variables in an expression.

val stmt_rhs : (Middle.Expr.Typed.t, 's) Middle.Stmt.Fixed.Pattern.t -> Middle.Expr.Typed.t Core_kernel.Set.Poly.t

The set of variables that can affect the value or behavior of the expression, i.e. rhs. Using Set.Poly instead of ExprSet so that 'e can be polymorphic, it usually doesn't matter if there's duplication.

val union_map : 'a Core_kernel.Set.Poly.t -> f:('a -> 'b Core_kernel.Set.Poly.t) -> 'b Core_kernel.Set.Poly.t

This is a helper function equivalent to List.concat_map but for Sets

val stmt_rhs_var_set : (Middle.Expr.Typed.t, 's) Middle.Stmt.Fixed.Pattern.t -> (Analysis_and_optimization.Dataflow_types.vexpr * Middle.Expr.Typed.Meta.t) Core_kernel.Set.Poly.t

The set of variables in an expression, including inside an index. For use in RHS sets, not LHS assignment sets, except in a target term.

val expr_assigned_var : Middle.Expr.Typed.t -> Analysis_and_optimization.Dataflow_types.vexpr

The variable being assigned to when the expression is the LHS

val summation_terms : Middle.Expr.Typed.t -> Middle.Expr.Typed.t list

The list of terms in expression separated by a +

val stmt_of_block : Middle.Stmt.Located.t list -> Middle.Stmt.Located.t

Represent a list of statements as a single statement

val subst_expr : (string, Middle.Expr.Typed.t) Core_kernel.Map.Poly.t -> Middle.Expr.Typed.t -> Middle.Expr.Typed.t

Substitute variables in an expression according to the provided Map.

val subst_stmt_base : (string, Middle.Expr.Typed.t) Core_kernel.Map.Poly.t -> (Middle.Expr.Typed.t, 'a) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, 'a) Middle.Stmt.Fixed.Pattern.t

Substitute variables occurring at the top level in statements according to the provided Map.

val subst_stmt : (string, Middle.Expr.Typed.t) Core_kernel.Map.Poly.t -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t

Substitute variables occurring anywhere in a statement according to the provided Map.

val name_subst_stmt : (string, string) Core_kernel.Map.Poly.t -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t

Substitute subexpressions occurring anywhere in a statement according to the provided Map.

val expr_subst_expr : Middle.Expr.Typed.t Middle.Expr.Typed.Map.t -> Middle.Expr.Typed.t -> Middle.Expr.Typed.t

Substitute subexpressions in an expression according to the provided Map, trying to match on larger subexpressions before smaller ones.

val expr_subst_stmt : Middle.Expr.Typed.t Middle.Expr.Typed.Map.t -> Middle.Stmt.Located.t -> Middle.Stmt.Located.t

Substitute subexpressions occurring anywhere in a statement according to the provided Map.

val expr_subst_stmt_base : Middle.Expr.Typed.t Middle.Expr.Typed.Map.t -> (Middle.Expr.Typed.t, 'a) Middle.Stmt.Fixed.Pattern.t -> (Middle.Expr.Typed.t, 'a) Middle.Stmt.Fixed.Pattern.t

Substitute subexpressions occurring at the top level in statements according to the provided Map.

val expr_depth : Middle.Expr.Typed.t -> int

Calculate how deeply nested an expression is.

val update_expr_ad_levels : string Core_kernel.Set.Poly.t -> Middle.Expr.Typed.t -> Middle.Expr.Typed.t

Recompute all AD-levels in the metadata of an expression from the bottom up, making the variables in the first argument autodiffable

val cleanup_empty_stmts : ('e, 's) Middle.Stmt.Fixed.t list -> ('e, 's) Middle.Stmt.Fixed.t list