Module Common.Nonempty_list

A non-empty list type. This can be constructed with the usual list syntax when the compiler is inferring the type, e.g.

let x : _ Nonempty_list.t = [1;2;3]

In Core v0.18, we can replace this with Core.Nonempty_list

type 'a t =

| :: of 'a * 'a list

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0__.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0__.Sexp.t) -> 'a t -> Sexplib0__.Sexp.t

val to_list : 'a t -> 'a list

val of_list_exn : 'a list -> 'a t