Module Common__.Helpers

type mem_pattern =

| AoS

| SoA

* This type represents whether or not an autodiff type can be represented * as an Array of Structs (AoS) or as a Struct of Arrays. This applies to * matrices, vectors, row vectors, and arrays of those types. * In the C++ this allows us to swap out matrix types from an * Eigen::Matrix<stan::math::var_value<double>, Rows, Cols> to an * stan::math::var_value<Eigen::Matrix<double, Rows, Cols>>. * (fyi a var in the C++ code is an alias for var_value<double>) * *

val map_mem_pattern : mem_pattern -> mem_pattern

val fold_mem_pattern : 'a -> mem_pattern -> 'a

val mem_pattern_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> mem_pattern

val sexp_of_mem_pattern : mem_pattern -> Ppx_sexp_conv_lib.Sexp.t

val compare_mem_pattern : mem_pattern -> mem_pattern -> int

val hash_fold_mem_pattern : Ppx_hash_lib.Std.Hash.state -> mem_pattern -> Ppx_hash_lib.Std.Hash.state

val hash_mem_pattern : mem_pattern -> Ppx_hash_lib.Std.Hash.hash_value

val lub_mem_pat : mem_pattern list -> mem_pattern

val option_or_else : if_none:'a option -> 'a option -> 'a option

val on_snd : ('a -> 'b) -> ('c * 'a) -> 'c * 'b

val on_fst : ('a -> 'b) -> ('a * 'c) -> 'b * 'c

val curry : (('a * 'b) -> 'c) -> 'a -> 'b -> 'c

val uncurry : ('a -> 'b -> 'c) -> ('a * 'b) -> 'c

val pp_builtin_syntax : string Fmt.t

val pp_keyword : string Fmt.t

val pp_angle_brackets : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit

val pp_brackets_postfix : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit