Module Frontend.Canonicalize

Re-write ASTs to remove deprecated features and clean up extra parenthesis, etc

type canonicalizer_settings = {

deprecations : bool;

parentheses : bool;

braces : bool;

inline_includes : bool;

strip_comments : bool;

}

Flags to enable or disable certain kinds of canonicalization. NB: inline_includes is controlled by the --canonicalize argument to stanc, but is consumed by the pretty-printer, not this module.

val legacy : canonicalizer_settings

Equivalent to what --print-canonical did before these settings were available

val none : canonicalizer_settings

val canonicalize_program : Ast.typed_program -> canonicalizer_settings -> Ast.typed_program

"Canonicalize" the program by removing deprecations, adding or removing parenthesis and braces, etc.