Module Frontend.Deprecation_removals

Generate errors for expired deprecations

Code in this module should generally only be hot for ONE version after the removal, then the errors will be moved to the parser/typechecker/etc as appropriate.

val collect_removed_expr : (Middle.Location_span.t * string) Base__List.t -> (Ast.typed_expr_meta, Ast.fun_kind) Ast.expr_with -> (Middle.Location_span.t * string) Base__List.t

val collect_removed_lval : (Middle.Location_span.t * string) Base__List.t -> Ast.typed_lval -> (Middle.Location_span.t * string) Base__List.t

val collect_removed_lval_pack : (Middle.Location_span.t * string) Base__List.t -> Ast.typed_lval Ast.lvalue_pack -> (Middle.Location_span.t * string) Base__List.t

val collect_removed_stmt : (Middle.Location_span.t * string) Base__List.t -> (Ast.typed_expression, Ast.stmt_typed_located_meta, Ast.typed_lval, Ast.fun_kind) Ast.statement_with -> (Middle.Location_span.t * string) Base__List.t

val collect_removals : Ast.typed_program -> (Middle.Location_span.t * string) Base__List.t

val pp : ?printed_filename:string -> Stdlib.Format.formatter -> (Middle.Location_span.t * string) -> unit

val pp_removals : ?printed_filename:string -> Stdlib.Format.formatter -> (Middle.Location_span.t * string) list -> unit