Module Frontend.Info

Implementation of the --info option which outputs on the standard output a JSON object containing information about the model.

The JSON object has a field inputs, parameters, transformed parameters, and generated quantities containing a dictionary where each entry corresponds to a variable in respectively the data, parameters, transformed parameters, and generated quantities blocks. To each variable is associated an object with two fields:

type: the base type of the variable ("int" or "real").

dimensions: the number of dimensions (0 for a scalar, 1 for a vector or row vector, etc.).

The JSON object also have the fields stanlib_calls and distributions containing the name of the standard library functions called and distributions used.

val info : Ast.typed_program -> string