Module Frontend.Semantic_error

type t

val pp : Stdlib.Format.formatter -> t -> unit

val location : t -> Middle.Location_span.t

val invalid_return : Middle.Location_span.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val mismatched_array_types : Middle.Location_span.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val invalid_row_vector_types : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val invalid_matrix_types : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val int_expected : Middle.Location_span.t -> string -> Middle.UnsizedType.t -> t

val int_or_real_expected : Middle.Location_span.t -> string -> Middle.UnsizedType.t -> t

val tuple_expected : Middle.Location_span.t -> string -> Middle.UnsizedType.t -> t

val int_intarray_or_range_expected : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val int_or_real_container_expected : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val scalar_or_type_expected : Middle.Location_span.t -> string -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val array_vector_rowvector_matrix_expected : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val illtyped_assignment : Middle.Location_span.t -> Middle.Operator.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val illtyped_ternary_if : Middle.Location_span.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val returning_fn_expected_nonreturning_found : Middle.Location_span.t -> string -> t

val returning_fn_expected_nonfn_found : Middle.Location_span.t -> string -> t

val returning_fn_expected_undeclaredident_found : Middle.Location_span.t -> string -> string option -> t

val returning_fn_expected_undeclared_dist_suffix_found : Middle.Location_span.t -> (string * string) -> t

val returning_fn_expected_wrong_dist_suffix_found : Middle.Location_span.t -> (string * string) -> t

val illtyped_reduce_sum_not_array : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val illtyped_reduce_sum_slice : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val illtyped_reduce_sum : Middle.Location_span.t -> string -> Middle.UnsizedType.t list -> Middle.UnsizedType.argumentlist -> SignatureMismatch.function_mismatch -> t

val ambiguous_function_promotion : Middle.Location_span.t -> string -> Middle.UnsizedType.t list option -> (Middle.UnsizedType.returntype * Middle.UnsizedType.argumentlist) list -> t

val illtyped_variadic : Middle.Location_span.t -> string -> Middle.UnsizedType.t list -> Middle.UnsizedType.argumentlist -> Middle.UnsizedType.t -> SignatureMismatch.function_mismatch -> t

val forwarded_function_signature_error : Middle.Location_span.t -> string -> string -> SignatureMismatch.details -> t

val forwarded_function_application_error : Middle.Location_span.t -> string -> string -> string list -> SignatureMismatch.details -> t

val illtyped_laplace_helper_args : Middle.Location_span.t -> string -> Middle.UnsizedType.argumentlist -> SignatureMismatch.details -> t

val illtyped_laplace_generic : Middle.Location_span.t -> string -> bool -> Middle.UnsizedType.argumentlist -> t

Generic failure. This means too few arguments were supplied, or that the function arguments are misplaced, both of which prevent us from giving a better message

val laplace_compatibility : Middle.Location_span.t -> string -> t

val illtyped_laplace_extra_args : Middle.Location_span.t -> string -> int -> t

val illtyped_laplace_tolerance_args : Middle.Location_span.t -> string -> SignatureMismatch.function_mismatch -> t

val nonreturning_fn_expected_returning_found : Middle.Location_span.t -> string -> t

val nonreturning_fn_expected_nonfn_found : Middle.Location_span.t -> string -> t

val nonreturning_fn_expected_undeclaredident_found : Middle.Location_span.t -> string -> string option -> t

val illtyped_fn_app : Middle.Location_span.t -> string -> (SignatureMismatch.signature_error list * bool) -> Middle.UnsizedType.t list -> t

val illtyped_binary_op : Middle.Location_span.t -> Middle.Operator.t -> Middle.UnsizedType.t -> Middle.UnsizedType.t -> t

val illtyped_prefix_op : Middle.Location_span.t -> Middle.Operator.t -> Middle.UnsizedType.t -> t

val illtyped_postfix_op : Middle.Location_span.t -> Middle.Operator.t -> Middle.UnsizedType.t -> t

val tuple_index_invalid_index : Middle.Location_span.t -> int -> int -> t

val tuple_index_not_tuple : Middle.Location_span.t -> Middle.UnsizedType.t -> t

val not_indexable : Middle.Location_span.t -> Middle.UnsizedType.t -> int -> t

val ident_is_keyword : Middle.Location_span.t -> string -> t

val ident_is_model_name : Middle.Location_span.t -> string -> t

val ident_is_stanmath_name : Middle.Location_span.t -> string -> t

val ident_in_use : Middle.Location_span.t -> string -> t

val ident_not_in_scope : Middle.Location_span.t -> string -> string option -> t

val invalid_decl_rng_fn : Middle.Location_span.t -> t

val invalid_rng_fn : Middle.Location_span.t -> t

val invalid_unnormalized_fn : Middle.Location_span.t -> t

val udf_is_unnormalized_fn : Middle.Location_span.t -> string -> t

val ident_has_unnormalized_suffix : Middle.Location_span.t -> string -> t

val conditional_notation_not_allowed : Middle.Location_span.t -> t

val conditioning_required : Middle.Location_span.t -> t

val not_printable : Middle.Location_span.t -> t

val empty_array : Middle.Location_span.t -> t

val empty_tuple : Middle.Location_span.t -> t

val bad_int_literal : Middle.Location_span.t -> t

val cannot_assign_to_read_only : Middle.Location_span.t -> string -> t

val cannot_assign_to_global : Middle.Location_span.t -> string -> t

val cannot_assign_function : Middle.Location_span.t -> string -> Middle.UnsizedType.t -> t

val cannot_assign_to_multiindex : Middle.Location_span.t -> t

val cannot_assign_duplicate_unpacking : Middle.Location_span.t -> Ast.untyped_lval list -> t

val cannot_access_assigning_var : Middle.Location_span.t -> string list -> t

val invalid_tilde_pdf_or_pmf : Middle.Location_span.t -> t

val invalid_tilde_cdf_or_ccdf : Middle.Location_span.t -> string -> t

val invalid_tilde_no_such_dist : Middle.Location_span.t -> string -> bool -> t

val target_plusequals_outside_model_or_logprob : Middle.Location_span.t -> t

val jacobian_plusequals_not_allowed : Middle.Location_span.t -> t

val invalid_truncation_cdf_or_ccdf : Middle.Location_span.t -> Middle.UnsizedType.argumentlist -> t

val break_outside_loop : Middle.Location_span.t -> t

val continue_outside_loop : Middle.Location_span.t -> t

val expression_return_outside_returning_fn : Middle.Location_span.t -> t

val void_outside_nonreturning_fn : Middle.Location_span.t -> t

val non_data_variable_size_decl : Middle.Location_span.t -> t

val non_int_bounds : Middle.Location_span.t -> t

val complex_transform : Middle.Location_span.t -> t

val transformed_params_int : Middle.Location_span.t -> t

val fn_overload_rt_only : Middle.Location_span.t -> string -> Middle.UnsizedType.returntype -> Middle.UnsizedType.returntype -> t

val fn_decl_redefined : Middle.Location_span.t -> string -> stan_math:bool -> Middle.UnsizedType.t -> t

val fn_decl_exists : Middle.Location_span.t -> string -> t

val fn_decl_without_def : Middle.Location_span.t -> string -> t

val fn_decl_needs_block : Middle.Location_span.t -> t

val non_real_prob_fn_def : Middle.Location_span.t -> Middle.UnsizedType.returntype -> t

val prob_density_non_real_variate : Middle.Location_span.t -> Middle.UnsizedType.t option -> t

val prob_mass_non_int_variate : Middle.Location_span.t -> Middle.UnsizedType.t option -> t

val duplicate_arg_names : Middle.Location_span.t -> t

val incompatible_return_types : Middle.Location_span.t -> t