Module Frontend__Environment

Type environments used during typechecking. Maps from strings to function or variable information

type originblock =

| MathLibrary

| Functions

| Data

| TData

| Param

| TParam

| Model

| GQuant

Origin blocks, to keep track of where variables are declared

val sexp_of_originblock : originblock -> Ppx_sexp_conv_lib.Sexp.t

val originblock_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> originblock

type varinfo = {

origin : originblock;

global : bool;

readonly : bool;

}

Information available for each variable

val sexp_of_varinfo : varinfo -> Ppx_sexp_conv_lib.Sexp.t

val varinfo_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> varinfo

type info = {

type_ : Middle.UnsizedType.t;

kind : [ `Variable of varinfo | `UserDeclared of Middle.Location_span.t | `StanMath | `UserDefined ];

}

val sexp_of_info : info -> Ppx_sexp_conv_lib.Sexp.t

val info_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> info

type t

val stan_math_environment : t

A type environment which contains the Stan math library functions

val find : t -> string -> info list

val add : t -> string -> Middle.UnsizedType.t -> [ `UserDeclared of Middle.Location_span.t | `StanMath | `UserDefined | `Variable of varinfo ] -> t

Add a new item to the type environment. Does not overwrite existing, but shadows

val set_raw : t -> string -> info list -> t

Overwrite the existing items bound to a name

val mem : t -> string -> bool

val iter : t -> (info list -> unit) -> unit

val nearest_ident : t -> string -> string option

The nearest identifier by edit distance, capped at edit distance 3 (if one exists)