Module Frontend__Parse

Some complicated stuff to get the custom syntax errors out of Menhir's Incremental API

val parse_file : (Stdlib.Lexing.position -> Frontend.Ast.untyped_program Frontend.Parser.MenhirInterpreter.checkpoint) -> string -> (Frontend.Ast.untyped_program, Frontend.Errors.t) Core_kernel.result * Frontend.Warnings.t list

A helper function to take a parser, a filename and produce an AST. Under the hood, it takes care of Menhir's custom syntax error messages.

val parse_string : (Stdlib.Lexing.position -> Frontend.Ast.untyped_program Frontend.Parser.MenhirInterpreter.checkpoint) -> string -> (Frontend.Ast.untyped_program, Frontend.Errors.t) Core_kernel.result * Frontend.Warnings.t list