Module Frontend__Promotion

module UnsizedType = Middle.UnsizedType

type t =

| NoPromotion

| IntToReal

| ToVar

| ToComplexVar

| IntToComplex

| RealToComplex

Type to represent promotions in the typechecker. This can be used to return information about how to promote expressions for use in Ast.Promotion

val promote_inner : Frontend.Ast.typed_expression -> t -> (Frontend.Ast.typed_expr_meta, Frontend.Ast.fun_kind) Frontend.Ast.expr_with

val promote : Frontend.Ast.typed_expression -> t -> (Frontend.Ast.typed_expr_meta, Frontend.Ast.fun_kind) Frontend.Ast.expr_with

val promote_list : Frontend.Ast.typed_expression list -> t list -> (Frontend.Ast.typed_expr_meta, Frontend.Ast.fun_kind) Frontend.Ast.expr_with list

val get_type_promotion_exn : ('a * UnsizedType.t) -> ('a * UnsizedType.t) -> t

Get the promotion needed to make the second type into the first. Types NEED to have previously been checked to be promotable

val promotion_cost : t -> int

Calculate the "cost"/number of promotions performed. Used to disambiguate function signatures