Module Middle.Expr

MIR types and modules corresponding to the expressions of the language

module Pattern : sig ... end

type 'a t = {

pattern : 'a t Pattern.t;

meta : 'a;

}

The "two-level" type for statements in the MIR. This corresponds to what the AST calls Frontend.Ast.expr_with

include Ppx_compare_lib.Comparable.S1 with type 'a t := 'a t

val compare : 'a Base__Ppx_compare_lib.compare -> 'a t Base__Ppx_compare_lib.compare

include Ppx_hash_lib.Hashable.S1 with type 'a t := 'a t

val hash_fold_t : 'a Base__Ppx_hash_lib.hash_fold -> 'a t Base__Ppx_hash_lib.hash_fold

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0__.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0__.Sexp.t) -> 'a t -> Sexplib0__.Sexp.t

val pp : 'a t Fmt.t

val rewrite_bottom_up : f:('a t -> 'a t) -> 'a t -> 'a t

rewrite_bottom_up specializes fold so that the result type 'r is equal to the type of our fixed-point data structure i.e. 'r = 'a t. This also means that the function f can be written with our fixed-point type 'a t as its argument.

module Typed : sig ... end

module Helpers : sig ... end