Module Middle__Index

Types of indexing operations

type 'a t =

| All

| Single of 'a

| Upfrom of 'a

| Between of 'a * 'a

| MultiIndex of 'a

val map : ('a -> 'b) -> 'a t -> 'b t

val fold : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

val t_of_sexp : a. (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

val sexp_of_t : a. ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_t : a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a t -> Ppx_hash_lib.Std.Hash.state

val compare : a. ('a -> 'a -> int) -> 'a t -> 'a t -> int

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

val pp_indexed : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> (string * 'a t list) -> unit

val bounds : 'a t -> 'a list

val apply : default:'b -> merge:('b -> 'b -> 'b) -> ('a -> 'b) -> 'a t -> 'b

Apply an op over the Index types inner expressions.

parameter default

Value to return for All

parameter merge

Function taking in lhs and rhs of Between and merging their result.

parameter op

a functor to run with inputs of inner exprs

parameter ind

the Index.t to

val folder : string Core_kernel.Set.Poly.t -> (string Core_kernel.Set.Poly.t -> 'a -> string Core_kernel.Set.Poly.t) -> 'a t -> string Core_kernel.Set.Poly.t