Module Middle__Stan_math_signatures

type dimensionality =

| DInt

| DReal

| DVector

| DMatrix

| DIntArray

| DVInt

| DVReal

| DIntAndReals

| DVectors

| DDeepVectorized

The "dimensionality" (bad name?) is supposed to help us represent the vectorized nature of many Stan functions. It allows us to represent when a function argument can be just a real or matrix, or some common forms of vectorization over reals. This captures the most commonly used forms in our previous signatures; there are a lot partially because we had a lot of inconsistencies.

val bare_array_type : (Middle.UnsizedType.t * int) -> Middle.UnsizedType.t

val expand_arg : dimensionality -> Middle.UnsizedType.t Base__List.t

type fkind =

| Lpmf

| Lpdf

| Rng

| Cdf

| Ccdf

| UnaryVectorized

val pp_fkind : Ppx_deriving_runtime.Format.formatter -> fkind -> Ppx_deriving_runtime.unit

val show_fkind : fkind -> Ppx_deriving_runtime.string

type fun_arg = Middle.UnsizedType.autodifftype * Middle.UnsizedType.t

type stan_math_table_values = Middle.UnsizedType.returntype * fun_arg list * Common.Helpers.mem_pattern

val is_primitive : Middle.UnsizedType.t -> bool

val stan_math_signatures : (string, stan_math_table_values list) Core_kernel.Hashtbl.t

The signatures hash table

val manual_stan_math_signatures : (string, stan_math_table_values list) Core_kernel.Hashtbl.t

All of the signatures that are added by hand, rather than the ones added "declaratively"

val all_combinations : 'a list list -> 'a list list

val missing_math_functions : Core_kernel.String.Set.t

val rng_return_type : Middle.UnsizedType.t -> Middle.UnsizedType.t list -> Middle.UnsizedType.t

val add_unqualified : (string Core_kernel.Hashtbl.key * Middle.UnsizedType.returntype * Middle.UnsizedType.t list * Common.Helpers.mem_pattern) -> unit

val ints_to_real : Middle.UnsizedType.t -> Middle.UnsizedType.t

val reduce_sum_allowed_dimensionalities : int list

val reduce_sum_slice_types : Middle.UnsizedType.t list

val variadic_ode_adjoint_ctl_tol_arg_types : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_ode_tol_arg_types : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_ode_mandatory_arg_types : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_ode_mandatory_fun_args : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_ode_fun_return_type : Middle.UnsizedType.t

val variadic_ode_return_type : Middle.UnsizedType.t

val variadic_dae_tol_arg_types : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_dae_mandatory_arg_types : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_dae_mandatory_fun_args : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list

val variadic_dae_fun_return_type : Middle.UnsizedType.t

val variadic_dae_return_type : Middle.UnsizedType.t

val mk_declarative_sig : (fkind list * string * dimensionality list * 'a) -> (Core_kernel.String.Set.Elt.t * Middle.UnsizedType.returntype * (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list * 'a) list

val full_lpdf : fkind list

val full_lpmf : fkind list

val reduce_sum_functions : Core_kernel.String.Set.t

val variadic_ode_adjoint_fn : string

val variadic_ode_nonadjoint_fns : Core_kernel.String.Set.t

val ode_tolerances_suffix : string

val is_reduce_sum_fn : Core_kernel.String.Set.Elt.t -> Core_kernel__.Import.bool

val is_variadic_ode_nonadjoint_fn : Core_kernel.String.Set.Elt.t -> Core_kernel__.Import.bool

val is_variadic_ode_fn : Core_kernel.String.Set.Elt.t -> bool

val is_variadic_ode_nonadjoint_tol_fn : Core_kernel.String.Set.Elt.t -> bool

val variadic_dae_fns : Core_kernel.String.Set.t

val dae_tolerances_suffix : string

val is_variadic_dae_fn : Core_kernel.String.Set.Elt.t -> Core_kernel__.Import.bool

val is_variadic_dae_tol_fn : Core_kernel.String.Set.Elt.t -> bool

val distributions : (fkind list * string * dimensionality list * Common.Helpers.mem_pattern) list

val math_sigs : (fkind list * string * dimensionality list * Common.Helpers.mem_pattern) list

val all_declarative_sigs : (fkind list * string * dimensionality list * Common.Helpers.mem_pattern) Base__List.t

val declarative_fnsigs : (Core_kernel.String.Set.Elt.t * Middle.UnsizedType.returntype * (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list * Common.Helpers.mem_pattern) list

val is_stan_math_function_name : string Core_kernel.Hashtbl.key -> bool

val is_soa_supported : string Core_kernel.Hashtbl.key -> fun_arg list -> bool

val dist_name_suffix : (string * 'a) list -> string -> string

val operator_to_stan_math_fns : Middle.Operator.t -> string list

val int_divide_type : Middle.UnsizedType.returntype * (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list * Common.Helpers.mem_pattern

val get_sigs : string Core_kernel.Hashtbl.key -> stan_math_table_values list

val make_assignmentoperator_stan_math_signatures : Middle.Operator.t -> (Middle.UnsizedType.returntype * (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list * Common.Helpers.mem_pattern) list

val pp_math_sig : Stdlib.Format.formatter -> (Middle.UnsizedType.returntype * (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t) list * Common.Helpers.mem_pattern) -> unit

val pp_math_sigs : Stdlib.Format.formatter -> string Core_kernel.Hashtbl.key -> unit

val pretty_print_math_sigs : string Core_kernel.Hashtbl.key -> string

val string_operator_to_stan_math_fns : string -> string

val pretty_print_all_math_sigs : Stdlib.Format.formatter -> unit -> unit

val pretty_print_all_math_distributions : Stdlib.Format.formatter -> unit -> unit

val pretty_print_math_lib_operator_sigs : Middle.Operator.t -> string list

val bare_types : Middle.UnsizedType.t list

val bare_types_size : int

val vector_types : Middle.UnsizedType.t list

val vector_types_size : int

val primitive_types : Middle.UnsizedType.t list

val primitive_types_size : int

val all_vector_types : Middle.UnsizedType.t list

val all_vector_types_size : int

val add_qualified : (string Core_kernel.Hashtbl.key * Middle.UnsizedType.returntype * fun_arg list * Common.Helpers.mem_pattern) -> unit

val add_nullary : string Core_kernel.Hashtbl.key -> unit

val add_binary : string Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_binary_vec : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_binary_vec_real_real : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_binary_vec_int_real : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_binary_vec_real_int : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_binary_vec_int_int : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_ternary : string Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val add_ternary_vec : string Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key Core_kernel.Hashtbl.key -> Common.Helpers.mem_pattern -> unit

val for_all_vector_types : (Middle.UnsizedType.t -> unit) -> unit

val for_vector_types : (Middle.UnsizedType.t -> unit) -> unit