Module Cpp.Stmts

Helpers for common statement constructs

val block : stmt list -> stmt

Wrap the list of statements in a block if it isn't a singleton block already

val unblock : stmt list -> stmt list

val rethrow_located : stmt list -> stmt list

Set up the try/catch logic for throwing an exception with its location set to the Stan program location.

val fori : identifier -> expr -> expr -> stmt -> stmt

val if_block : expr -> stmt list -> stmt

val unused : identifier -> stmt list

Suppress warnings for a variable which may not be used.