Module Stan_math_backend.Lower_stmt

Lowering of Stan statements to C++

val lower_st : 'a Middle.SizedType.t -> Middle.UnsizedType.autodifftype -> Cpp.type_

val check_to_string : 'a Middle.Transformation.t -> string option

val math_fn_translations : 'a Middle.Internal_fun.t -> string option

val trans_math_fn : 'a Middle.Internal_fun.t -> string

val initialize_value : Middle.Expr.Typed.t Middle.SizedType.t -> Middle.UnsizedType.autodifftype -> Cpp.expr

Code generation for the right hand side of expressions to initialize objects. For scalar types this sets the value to NaN and for containers initializes the memory.

val lower_assign_sized : Middle.Expr.Typed.t Middle.SizedType.t -> Middle.UnsizedType.autodifftype -> Middle.Expr.Typed.t Middle.Stmt.Pattern.decl_init -> Cpp.expr option

Initialize an object of a given size.

val lower_unsized_decl : Cpp.identifier -> Middle.UnsizedType.t -> Middle.UnsizedType.autodifftype -> Cpp.variable_defn

val lower_possibly_opencl_decl : string -> 'a Middle.SizedType.t -> Middle.UnsizedType.autodifftype -> 'b Middle.Expr.t Middle.Stmt.Pattern.decl_init -> Cpp.type_

val lower_sized_decl : Cpp.identifier -> Middle.Expr.Typed.t Middle.SizedType.t -> Middle.UnsizedType.autodifftype -> Middle.Expr.Typed.Meta.t Middle.Expr.t Middle.Stmt.Pattern.decl_init -> Cpp.variable_defn

val lower_decl : Cpp.identifier -> Middle.Expr.Typed.t Middle.Type.t -> Middle.UnsizedType.autodifftype -> Middle.Expr.Typed.Meta.t Middle.Expr.t Middle.Stmt.Pattern.decl_init -> Cpp.stmt

val lower_profile : Cpp.identifier -> Cpp.stmt list -> Stan_math_backend__Cpp.stmt

val lower_bool_expr : Middle.Expr.Typed.t -> Cpp.expr

val lower_nonrange_lvalue : Middle.Expr.Typed.t Middle.Stmt.Pattern.lvalue -> Cpp.expr

val lower_nonrange_lbase : Middle.Expr.Typed.t Middle.Stmt.Pattern.lbase -> Cpp.expr

val expr_overlaps_lhs_ref : 'a Middle.Expr.t Middle.Stmt.Pattern.lvalue -> 'a Middle.Expr.t -> bool

val throw_exn : string -> Middle.Expr.Typed.t list -> Stan_math_backend__Cpp.stmt

val lower_statement : (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.t -> Cpp.stmt list

val lower_statements : (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.t list -> Cpp.stmt Base__List.t

module Testing : sig ... end