Module Stan_math_backend.Statement_gen

val pp_call_str : Stdlib.Format.formatter -> (string * string list) -> unit

val pp_block : Stdlib.Format.formatter -> ((Stdlib.Format.formatter -> 'a -> unit) * 'a) -> unit

val pp_profile : Stdlib.Format.formatter -> ((Stdlib.Format.formatter -> 'a -> unit) * string * 'a) -> unit

val contains_eigen : Middle.UnsizedType.t -> bool

val pp_filler : Stdlib.Format.formatter -> (string * 'a Middle.SizedType.t * string * bool) -> unit

val pp_st : Stdlib.Format.formatter -> ('a Middle.SizedType.t * Middle.UnsizedType.autodifftype) -> unit

val pp_ut : Stdlib.Format.formatter -> (Middle.UnsizedType.t * Middle.UnsizedType.autodifftype) -> unit

val nan_type : ('a * Middle.UnsizedType.autodifftype) -> string

val pp_initialize : Stdlib.Format.formatter -> (Middle.Expr.Typed.t Middle.SizedType.t * Middle.UnsizedType.autodifftype) -> unit

val pp_assign_sized : Stdlib.Format.formatter -> (Middle.Expr.Typed.t Middle.SizedType.t * Middle.UnsizedType.autodifftype * bool) -> unit

val pp_assign_data : Stdlib.Format.formatter -> (string * Middle.Expr.Typed.t Middle.SizedType.t * bool) -> unit

Initialize Data and Transformed Data This function is used in the model's constructor to

Initialize memory for the data and transformed data

If an Eigen type, place that memory into the class's Map

Set the initial values of that data to NaN

parameter ppf

A pretty printer

parameter decl_id

The name of the model class member

parameter st

The type of the class member

val pp_for_loop : Stdlib.Format.formatter -> (string * Middle.Expr.Typed.t * Middle.Expr.Typed.t * (Stdlib.Format.formatter -> 'a -> unit) * 'a) -> unit

pp_for_loop ppf (loopvar, lower, upper, pp_body, body) tries to pretty print a for-loop from lower to upper given some loopvar.

val integer_el_type : 'a Middle.SizedType.t -> bool

val pp_data_decl : Stdlib.Format.formatter -> (string * Middle.UnsizedType.t) -> unit

Print the private members of the model class

Accounting for types that can be moved to OpenCL.

parameter ppf

A formatter

parameter vident

name of the private member.

parameter ut

The unsized type to print.

val pp_emplace_var : Stdlib.Format.formatter -> Middle.Expr.Typed.t -> unit

Create string representations for vars__.emplace_back

val pp_map_decl : Stdlib.Format.formatter -> (string * Middle.UnsizedType.t) -> unit

Create strings representing maps of Eigen types

val pp_unsized_decl : Stdlib.Format.formatter -> (string * Middle.UnsizedType.t * Middle.UnsizedType.autodifftype) -> unit

val pp_possibly_opencl_decl : Stdlib.Format.formatter -> (string * 'a Middle.SizedType.t * Middle.UnsizedType.autodifftype) -> unit

val pp_sized_decl : Stdlib.Format.formatter -> (string * Middle.Expr.Typed.t Middle.SizedType.t * Middle.UnsizedType.autodifftype * bool) -> unit

val pp_decl : Stdlib.Format.formatter -> (string * Middle.Expr.Typed.t Middle.Type.t * Middle.UnsizedType.autodifftype * bool) -> unit

val math_fn_translations : 'a Middle.Internal_fun.t -> (string * 'b list) option

val trans_math_fn : 'a Middle.Internal_fun.t -> string * 'b list

val pp_bool_expr : Stdlib.Format.formatter -> Middle.Expr.Typed.t -> unit

val pp_statement : (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t Fmt.t

val pp_block_s : Stdlib.Format.formatter -> (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t -> unit