Module Stan_math_backend__Cpp

A set of data types representing the C++ we generate

type identifier = string

val identifier_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> identifier

val sexp_of_identifier : identifier -> Ppx_sexp_conv_lib.Sexp.t

type type_ =

| Auto

| Void

| Int

| Double

| Complex of type_

| TemplateType of identifier

| StdVector of type_
A std::vector. For Eigen Vectors, use Matrix with a row or column size of 1

| Array of type_ * int

| TypeLiteral of identifier
Used for things like Eigen::Index

| Matrix of type_ * int * int

| Ref of type_

| Const of type_

| Pointer of type_

| TypeTrait of identifier * type_ list
e.g. stan::promote_scalar, stan:base_type

C++ types

val type__of_sexp : Ppx_sexp_conv_lib.Sexp.t -> type_

val sexp_of_type_ : type_ -> Ppx_sexp_conv_lib.Sexp.t

module Types : sig ... end

type operator =

| Multiply

| Divide

| Add

| Subtract

| Eq

| LEq

| GEq

| Lthn

| Gthn

| And

| Or

val operator_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> operator

val sexp_of_operator : operator -> Ppx_sexp_conv_lib.Sexp.t

type expr =

| Literal of string
printed as-is

| Var of identifier

| VarRef of identifier

| Parens of expr

| FunCall of identifier * type_ list * expr list

| MethodCall of expr * identifier * type_ list * expr list

| StaticMethodCall of type_ * identifier * type_ list * expr list

| Constructor of type_ * expr list
printed as type(expr1, expr2, ...)

| InitializerExpr of type_ * expr list
printed as type{expr1, expr2, ...}

| ArrayLiteral of expr list

| TernaryIf of expr * expr * expr

| Cast of type_ * expr

| Index of expr * expr

| AllocNew of type_ * expr list

| OperatorNew of identifier * type_ * expr list
See operator new for distinctions between allocating and placing news

| Assign of expr * expr
NB: Not all exprs are valid lvalues!

| StreamInsertion of expr * expr list
Corresponds to operator<<

| BinOp of expr * operator * expr

| PMinus of expr

| Increment of expr

val expr_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> expr

val sexp_of_expr : expr -> Ppx_sexp_conv_lib.Sexp.t

module Exprs : sig ... end

module Expression_syntax : sig ... end

type init =

| Assignment of expr

| Construction of expr list

| InitializerList of expr list

| Uninitialized

val init_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> init

val sexp_of_init : init -> Ppx_sexp_conv_lib.Sexp.t

type variable_defn = {

static : bool;

constexpr : bool;

type_ : type_;

name : identifier;

init : init;

}

val make_variable_defn : ?⁠static:bool -> ?⁠constexpr:bool -> type_:type_ -> name:identifier -> ?⁠init:init -> unit -> variable_defn

val variable_defn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> variable_defn

val sexp_of_variable_defn : variable_defn -> Ppx_sexp_conv_lib.Sexp.t

type stmt =

| Expression of expr

| VariableDefn of variable_defn

| For of variable_defn * expr * expr * stmt

| ForEach of type_ * identifier * expr * stmt

| While of expr * stmt

| IfElse of expr * stmt * stmt option

| TryCatch of stmt list * type_ * identifier * stmt list

| Block of stmt list

| Return of expr option

| Throw of expr

| Break

| Continue

| Semicolon

| Using of string * type_ option

| Comment of string

val stmt_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> stmt

val sexp_of_stmt : stmt -> Ppx_sexp_conv_lib.Sexp.t

module Stmts : sig ... end

module Decls : sig ... end

type template_parameter =

| Typename of string
The name of a template typename

| RequireIs of string * string
A C++ type trait (e.g. is_arithmetic) and the template name which needs to satisfy that. These are collated into one require_all_t<>

| Require of string * string list

| Bool of string
A named boolean template type

val template_parameter_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> template_parameter

val sexp_of_template_parameter : template_parameter -> Ppx_sexp_conv_lib.Sexp.t

type cv_qualifiers =

| Const

| Final

| NoExcept

val cv_qualifiers_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> cv_qualifiers

val sexp_of_cv_qualifiers : cv_qualifiers -> Ppx_sexp_conv_lib.Sexp.t

type fun_defn = {

templates_init : template_parameter list list * bool;
Double list since some functions (mainly reduce_sum functors) need two sets of templates

inline : bool;

return_type : type_;

name : identifier;

args : (type_ * string) list;

cv_qualifiers : cv_qualifiers list;

body : stmt list option;

}

val make_fun_defn : ?⁠templates_init:(template_parameter list list * bool) -> ?⁠inline:bool -> return_type:type_ -> name:identifier -> ?⁠args:(type_ * string) list -> ?⁠cv_qualifiers:cv_qualifiers list -> ?⁠body:stmt list -> unit -> fun_defn

val fun_defn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> fun_defn

val sexp_of_fun_defn : fun_defn -> Ppx_sexp_conv_lib.Sexp.t

type constructor = {

args : (type_ * string) list;

init_list : (identifier * expr list) list;

body : stmt list;

}

val make_constructor : ?⁠args:(type_ * string) list -> ?⁠init_list:(identifier * expr list) list -> ?⁠body:stmt list -> unit -> constructor

val constructor_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> constructor

val sexp_of_constructor : constructor -> Ppx_sexp_conv_lib.Sexp.t

type directive =

| Include of string

| IfNDef of string * defn list

| MacroApply of string * string list

Incomplete list of C++ preprocessor directives

and class_defn = {

class_name : identifier;

final : bool;

public_base : type_;

private_members : defn list;

public_members : defn list;

constructor : constructor;

destructor_body : stmt list;

}

The Stan model class always has a non-default constructor and destructor

and struct_defn = {

param : template_parameter option;

struct_name : identifier;

body : defn list;

}

and defn =

| FunDef of fun_defn

| Class of class_defn

| Struct of struct_defn

| GlobalVariableDefn of variable_defn

| GlobalComment of string

| GlobalUsing of string * type_ option

| Namespace of identifier * defn list

| Preprocessor of directive

val directive_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> directive

val class_defn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> class_defn

val struct_defn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> struct_defn

val defn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> defn

val sexp_of_directive : directive -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_class_defn : class_defn -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_struct_defn : struct_defn -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_defn : defn -> Ppx_sexp_conv_lib.Sexp.t

val make_class_defn : name:identifier -> public_base:type_ -> ?⁠final:bool -> private_members:defn list -> public_members:defn list -> constructor:constructor -> ?⁠destructor_body:stmt list -> unit -> class_defn

val make_struct_defn : param:template_parameter option -> name:identifier -> body:defn list -> unit -> struct_defn

type program = defn list

Much like in C++, we define a translation unit as a list of definitions

val program_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> program

val sexp_of_program : program -> Ppx_sexp_conv_lib.Sexp.t

module Printing : sig ... end

module Tests : sig ... end