Module Stan_math_backend__Numbering

This module translates the locations of statements into an array of messages used in the C++ for providing locations in the Stan code when an error occurs, and replaces the locations in the MIR tree with indices into said array.

It also numbers calls to map_rect for registration.

type state_t

type map_rect_registration_t

val prepare_prog : Middle.Program.Typed.t -> Middle.Program.Numbered.t * state_t * map_rect_registration_t

val no_span_num : Middle.Stmt.Numbered.Meta.t

val gen_globals : state_t -> Stan_math_backend.Cpp.defn list

val assign_loc : Middle.Stmt.Numbered.Meta.t -> Stan_math_backend.Cpp.stmt list

val register_map_rect_functors : string -> map_rect_registration_t -> Stan_math_backend.Cpp.defn list

Register functiors used for map_rect.