Module Stan_math_backend__Stan_math_code_gen

Generate C++ from the MIR.

This module makes extensive use of the Format0 module via the Fmt1 API. As such, you'll need to understand the "%a" and "@" notation from 0, especially the section headed "Formatted pretty-printing." Then, we use functions like pf and str from the Fmt library1. On top of that, the "@" pretty-printing specifiers all actually correspond 1-to-1 with commands like open_box from the Format library0. The boxing system is best described in this explainer pdf 2, particularly Section 3 ("Format basics"). It's worth noting that someone was able to make a decent-looking pretty-printer for a subset of Javascript3 that might serve as a good reference. Good luck!

0 Format module doc: https://caml.inria.fr/pub/docs/manual-ocaml/libref/Format.html 1 Fmt module doc: https://erratique.ch/software/fmt/doc/Fmt.html 2 Format Unraveled: https://hal.archives-ouvertes.fr/hal-01503081/file/format-unraveled.pdf 3 Javascript pretty-printer https://github.com/Virum/compiler/blob/28e807b842bab5dcf11460c8193dd5b16674951f/JavaScript.ml#L112

type template =

| Typename of string

| Require of string * string

| Bool of string

val pp_template : Stdlib.Format.formatter -> template -> unit

val pp_template_defaults : Stdlib.Format.formatter -> template -> unit

val pp_templates : defaults:bool -> Stdlib.Format.formatter -> template list -> unit

type found_functor = {

struct_template : template option;

arg_templates : template list;

signature : string;

defn : string;

}

val standalone_functions : bool Core_kernel.ref

val stanc_args_to_print : string

val pp_unused : Stdlib.Format.formatter -> string -> unit

val pp_function__ : Stdlib.Format.formatter -> (string * string) -> unit

Print name of model function.

parameter prog_name

Name of the Stan program.

parameter fname

Name of the function.

val pp_located : Stdlib.Format.formatter -> 'a -> unit

Print the body of exception handling for functions

val arg_needs_template : (Middle.UnsizedType.autodifftype * 'a * Middle.UnsizedType.t) -> bool

Detect if argument requires C++ template

val maybe_templated_arg_types : Middle.Program.fun_arg_decl -> string option list

Print template arguments for C++ functions that need templates

parameter args

A pack of Program.fun_arg_decl containing functions to detect templates.

returns

A list of arguments with template parameter names added.

val maybe_require_templates : string option list -> Middle.Program.fun_arg_decl -> template option list

val return_arg_types : Middle.Program.fun_arg_decl -> string option list

val pp_promoted_scalar : Stdlib.Format.formatter -> Middle.Program.fun_arg_decl -> unit

Print the code for promoting stan real types

parameter ppf

A pretty printer

parameter args

A pack of arguments to detect whether they need to use the promotion rules.

val pp_returntype : Stdlib.Format.formatter -> Middle.Program.fun_arg_decl -> Middle.UnsizedType.t option -> unit

Pretty-prints a function's return-type, taking into account templated argument promotion.

val pp_eigen_arg_to_ref : Stdlib.Format.formatter -> ('a * string * Middle.UnsizedType.t) list -> unit

val pp_located_error : Stdlib.Format.formatter -> ((Stdlib.Format.formatter -> 'a -> unit) * 'a) -> unit

pp_located_error ppf (pp_body_block, body_block, err_msg) surrounds body_block with a C++ try-catch that will rethrow the error with the proper source location from the body_block (required to be a stmt_loc Block variant).

parameter ppf

A pretty printer.

parameter pp_body_block

A pretty printer for the body block

parameter body

A C++ scoped body block surrounded by squiggly braces.

val pp_arg : Stdlib.Format.formatter -> (string option * ('a * string * Middle.UnsizedType.t)) -> unit

Print the type of an object.

parameter ppf

A pretty printer

parameter custom_scalar_opt

A string representing a types inner scalar value.

parameter name

The name of the object

parameter ut

The unsized type of the object

val pp_arg_eigen_suffix : Stdlib.Format.formatter -> (string option * ('a * string * Middle.UnsizedType.t)) -> unit

val pp_located_error_b : Stdlib.Format.formatter -> (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t list -> unit

pp_located_error_b automatically adds a Block wrapper

val typename : string -> template

val get_templates_and_args : bool -> Middle.Program.fun_arg_decl -> string list * template list * string list

Construct an object with it's needed templates for function signatures.

parameter fdargs

A sexp list of strings representing C++ types.

val pp_template_decorator : Stdlib.Format.formatter -> string list -> unit

Print the C++ template parameter decleration before a function.

parameter ppf

A pretty printer.

val mk_extra_args : string list -> string list -> string list

val pp_fun_def : Stdlib.Format.formatter -> ((Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t Middle.Program.fun_def * (string, found_functor list) Core_kernel.Hashtbl.t * (string * template list) Core_kernel.Hash_set.t * Core_kernel.String.Set.t * Core_kernel.String.Set.t * Core_kernel.String.Set.t) -> unit

Print the C++ function definition.

parameter ppf

A pretty printer Refactor this please - one idea might be to have different functions for printing user defined distributions vs rngs vs regular functions.

val pp_standalone_fun_def : string -> Stdlib.Format.formatter -> 'a Middle.Program.fun_def -> unit

Creates functions outside the model namespaces which only call the ones inside the namespaces

val version : string

val includes : string

val pp_validate_data : Stdlib.Format.formatter -> (string * Middle__Expr.Typed.t Middle.SizedType.t) -> unit

Validate the dimensions of the C++ object are correct at runtime

parameter ppf

A pretty printer

parameter name

The name of the object.

parameter st

The SizedType of the object.

val pp_mul : Stdlib.Format.formatter -> unit -> unit

val get_unconstrained_param_st : ('a * Middle__Expr.Typed.t Middle.Program.outvar) -> Middle__Expr.Typed.t list option

val get_constrained_param_st : ('a * Middle__Expr.Typed.t Middle.Program.outvar) -> Middle__Expr.Typed.t list option

val pp_num_param : Stdlib.Format.formatter -> Middle.Expr.Typed.t list -> unit

val pp_ctor : Stdlib.Format.formatter -> (Middle__Expr.Typed.t, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the constructor of the model class. Read in data steps: 1. context__.validate_dims() to verify the dimensions are correct at runtime. 1. find vals_%s__ from context__.vals_%s(vident) 1. keep track of pos__ 1. run checks on resulting vident

val top_level_decls : ('a, 'b) Middle.Stmt.Fixed.t -> (string * Middle.UnsizedType.t) list

val pp_model_private : Stdlib.Format.formatter -> ('a, ('b, 'c) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the private data members of the model class

val pp_method : Stdlib.Format.formatter -> string -> string -> string list -> (Stdlib.Format.formatter -> unit -> unit) -> ?⁠outro:unit Fmt.t -> ?⁠cv_attr:string list -> (Stdlib.Format.formatter -> unit) -> unit

Print the signature and blocks of the model class methods.

parameter ppf

A pretty printer

parameter rt

The return type.

parameter name

The method name.

parameter intro

Anything that needs printed before the method body.

parameter outro

Anything that needs printed after the method body.

parameter cv_attr

Optional parameter to add method attributes.

parameter ppbody

(?A pretty printer of the method's body)

val pp_get_param_names : Stdlib.Format.formatter -> ('a, 'b) Middle.Program.t -> unit

Print the get_param_names method of the model class

parameter ppf

A pretty printer.

val pp_get_dims : Stdlib.Format.formatter -> (Middle__Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the get_dims method of the model class.

val pp_method_b : Stdlib.Format.formatter -> string -> string -> string list -> (Stdlib.Format.formatter -> unit -> unit) -> ?⁠outro:unit Fmt.t -> ?⁠cv_attr:string list -> (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t list -> unit

val pp_write_array : Stdlib.Format.formatter -> ('a, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the write_array method of the model class

val pp_for_loop_iteratee : ?⁠index_ids:string list -> Stdlib.Format.formatter -> ('a * Middle.Expr.Typed.t list * (Stdlib.Format.formatter -> ('a * string list) -> unit)) -> unit

Prints the for loop for constrained_param_names and unconstrained_param_names

parameter index_ids

Optional named parameter of a SizedType's dimensions

parameter ppf

A pretty printer

parameter dims

A list of the dimensions of a SizedType

parameter pp_body

Pretty printer for the body of the loop.

val emit_name : Stdlib.Format.formatter -> (string * string list) -> unit

val emit_complex_name : Stdlib.Format.formatter -> (string * string list) -> unit

val pp_constrained_param_names : Stdlib.Format.formatter -> (Middle.Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the constrained_param_names method of the model class.

val pp_unconstrained_param_names : Stdlib.Format.formatter -> (Middle.Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the unconstrained_param_names method of the model class. This is just a copy of constrained, I need to figure out which one is wrong and fix it eventually. From Bob,

Off the top of my head, I think the four that change sizes from constrained to unconstrained are:

simplex: K -> (K - 1) covar_matrix: K^2 -> (K choose 2) + K corr_matrix: K^2 -> (K choose 2) cholesky_corr: (K choose 2) + K -> (K choose 2)

cholesky_cov does not change size (it's (K choose 2) + K). Now that our unit vector uses the normal thing, that also doesn't change size. The ordered types and constrained types don't change sizes either.

val pp_transform_inits_impl : Stdlib.Format.formatter -> ('a, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the transform_inits method of the model class

val pp_log_prob : Stdlib.Format.formatter -> ('a, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the log_prob method of the model class

val pp_outvar_metadata : Stdlib.Format.formatter -> (string * (string * Middle.Expr.Typed.t Middle.SizedType.t * Middle.Program.io_block) list) -> unit

Print the body of the constrained and unconstrained sizedtype methods in the model class

parameter ppf

A pretty printer

parameter method_name

The name of the method to wrap the body in.

parameter outvars

The parameters to gather the sizes for.

val pp_unconstrained_types : Stdlib.Format.formatter -> (Middle.Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the get_unconstrained_sizedtypes method of the model class

val pp_constrained_types : Stdlib.Format.formatter -> (Middle.Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the get_constrained_sizedtypes method of the model class

val pp_overloads : Stdlib.Format.formatter -> (Middle.Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the generic method overloads needed in the model class.

val pp_transform_inits : Stdlib.Format.formatter -> (Middle__Expr.Typed.t, 'a) Middle.Program.t -> unit

Print the get_constrained_sizedtypes method of the model class

val pp_model_public : Stdlib.Format.formatter -> (Middle__Expr.Typed.t, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the public parts of the model class

val model_prefix : string

val pp_model : Stdlib.Format.formatter -> (Middle__Expr.Typed.t, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> unit

Print the full model class.

val usings : string

val namespace : ('a, 'b) Middle.Program.t -> string

Create the model's namespace.

val pp_register_map_rect_functors : Stdlib.Format.formatter -> ('a, 'b) Middle.Program.t -> unit

Find and register functiors used for map_rect.

val is_fun_used_with_variadic_fn : (string -> bool) -> ('a Middle.Expr.Fixed.t, ('b, 'c) Middle.Stmt.Fixed.t) Middle.Program.t -> Core_kernel.String.Set.t

val collect_functors_functions : ('a Middle.Expr.Fixed.t, (Middle.Expr.Typed.Meta.t, Middle.Stmt.Numbered.Meta.t) Middle.Stmt.Fixed.t) Middle.Program.t -> string * (string, found_functor list) Core_kernel.Hashtbl.t

val pp_prog : Stdlib.Format.formatter -> Middle.Program.Typed.t -> unit

Print the full C++ for the stan program.