Module type Monotone_framework_sigs.FLOWGRAPH

The API for a flowgraph, needed for the mfp algorithm in the monotone framework. Assumed invariants: successors contains all graph nodes as keys initials is a subset of the graph nodes

type labels

type t = labels

val compare : t Base__Ppx_compare_lib.compare

val sexp_of_t : t -> Sexplib0.Sexp.t

val hash : t -> int

val initials : labels Core.Set.Poly.t

val successors : (labels, labels Core.Set.Poly.t) Core.Map.Poly.t