Module type Monotone_framework_sigs.FLOWGRAPH

The API for a flowgraph, needed for the mfp algorithm in the monotone framework. Assumed invariants: successors contains all graph nodes as keys initials is a subset of the graph nodes

type labels

include Base__.Hashtbl_intf.Key.S with type FLOWGRAPH.t = labels

type t = labels

val compare : t -> t -> int

val sexp_of_t : t -> Base__.Ppx_sexp_conv_lib.Sexp.t

val hash : t -> int

val initials : labels Core_kernel.Set.Poly.t

val successors : (labels, labels Core_kernel.Set.Poly.t) Core_kernel.Map.Poly.t