Module Common

module Gensym : sig ... end

Generate fresh (never before used) symbols.

module ICE : sig ... end

Internal compiler errors

module Let_syntax : sig ... end

See https://ocaml.org/manual/5.2/bindingops.html#ss%3Aletops-conventions This is an alternative to the let%bind and let%map syntax from ppx_let: https://blog.janestreet.com/let-syntax-and-why-you-should-use-it/

module Nonempty_list : sig ... end

A non-empty list type. This can be constructed with the usual list syntax when the compiler is inferring the type, e.g.