Module Common.Foldable

module type Basic = sig ... end

The Basic definition for type constructors with a single type variable.

module type S = sig ... end

The complete module signature for Foldable for types with a single type variable.

module Make = Foldable.Make

module type Basic2 = sig ... end

The Basic definition for type constructors with two type variables.

module type S2 = sig ... end

The complete module signature for Foldable for types with a single type variable.

module Make2 = Foldable.Make2