Module type Foldable.Basic

The Basic definition for type constructors with a single type variable.

This signature just says that you have some type 'a t and a function fold with type:

fold: ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b

This function can be derived using ppx_deriving.

type 'a t

val fold : ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b