Module Common.ICE

Internal compiler errors

val internal_compiler_error : Base__.Sexp.t -> 'a

An alias of Core.raise_s. This used to do more processing, for now it is preserved just as a nicer marker in the code

val with_exn_message : (unit -> 'a) -> ('a, string) Core._result

Catch all exceptions at the top-level and convert them into a 'a, string result where the string contains the exception and a backtrace if present, followed by a link to our bugtracker.