Module Common.Specialized

module type Meta = sig ... end

Signature of all meta data used to annotate IRs

module type Unspecialized = sig ... end

module type Unspecialized2 = sig ... end

module type S = sig ... end

Since the type t is now concrete (i.e. not a type _constructor_) we can construct a Comparable.S giving us Map and Set specialized to the type.

module Make = Specialized.Make

module Make2 = Specialized.Make2