Module Frontend.Ast

type identifier = {

name : string;

id_loc : Middle.Location_span.t;

}

Our type for identifiers, on which we record a location

val identifier_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> identifier

val sexp_of_identifier : identifier -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_identifier : Ppx_hash_lib.Std.Hash.state -> identifier -> Ppx_hash_lib.Std.Hash.state

val hash_identifier : identifier -> Ppx_hash_lib.Std.Hash.hash_value

val compare_identifier : identifier -> identifier -> int

type 'e index =

| All

| Single of 'e

| Upfrom of 'e

| Downfrom of 'e

| Between of 'e * 'e

Indices for array access

val map_index : ('a -> 'b) -> 'a index -> 'b index

val fold_index : ('a -> 'b -> 'a) -> 'a -> 'b index -> 'a

val index_of_sexp : e. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> Ppx_sexp_conv_lib.Sexp.t -> 'e index

val sexp_of_index : e. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> 'e index -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_index : e. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'e index -> Ppx_hash_lib.Std.Hash.state

val compare_index : e. ('e -> 'e -> int) -> 'e index -> 'e index -> int

type fun_kind =

| StanLib of bool Middle.Fun_kind.suffix

| UserDefined of bool Middle.Fun_kind.suffix

Front-end function kinds

val compare_fun_kind : fun_kind -> fun_kind -> int

val fun_kind_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> fun_kind

val sexp_of_fun_kind : fun_kind -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_fun_kind : Ppx_hash_lib.Std.Hash.state -> fun_kind -> Ppx_hash_lib.Std.Hash.state

val hash_fun_kind : fun_kind -> Ppx_hash_lib.Std.Hash.hash_value

type ('e, 'f) expression =

| TernaryIf of 'e * 'e * 'e

| BinOp of 'e * Middle.Operator.t * 'e

| PrefixOp of Middle.Operator.t * 'e

| PostfixOp of 'e * Middle.Operator.t

| Variable of identifier

| IntNumeral of string

| RealNumeral of string

| ImagNumeral of string

| FunApp of 'f * identifier * 'e list

| CondDistApp of 'f * identifier * 'e list

| Promotion of 'e * Middle.UnsizedType.t * Middle.UnsizedType.autodifftype

| GetLP

| GetTarget

| ArrayExpr of 'e list

| RowVectorExpr of 'e list

| Paren of 'e

| Indexed of 'e * 'e index list

Expression shapes (used for both typed and untyped expressions, where we substitute untyped_expression or typed_expression for 'e

val map_expression : ('a -> 'b) -> ('c -> 'd) -> ('a, 'c) expression -> ('b, 'd) expression

val fold_expression : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> 'a -> ('b, 'c) expression -> 'a

val expression_of_sexp : e f. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> (Ppx_sexp_conv_lib.Sexp.t -> 'f) -> Ppx_sexp_conv_lib.Sexp.t -> ('e, 'f) expression

val sexp_of_expression : e f. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> ('f -> Ppx_sexp_conv_lib.Sexp.t) -> ('e, 'f) expression -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_expression : e f. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'f -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('e, 'f) expression -> Ppx_hash_lib.Std.Hash.state

val compare_expression : e f. ('e -> 'e -> Core_kernel__.Import.int) -> ('f -> 'f -> Core_kernel__.Import.int) -> ('e, 'f) expression -> ('e, 'f) expression -> Core_kernel__.Import.int

type ('m, 'f) expr_with = {

expr : (('m, 'f) expr_with, 'f) expression;

emeta : 'm;

}

val map_expr_with : ('m -> 'a) -> ('f -> 'b) -> ('m, 'f) expr_with -> ('a, 'b) expr_with

val fold_expr_with : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> 'a -> ('b, 'c) expr_with -> 'a

val expr_with_of_sexp : m f. (Ppx_sexp_conv_lib.Sexp.t -> 'm) -> (Ppx_sexp_conv_lib.Sexp.t -> 'f) -> Ppx_sexp_conv_lib.Sexp.t -> ('m, 'f) expr_with

val sexp_of_expr_with : m f. ('m -> Ppx_sexp_conv_lib.Sexp.t) -> ('f -> Ppx_sexp_conv_lib.Sexp.t) -> ('m, 'f) expr_with -> Ppx_sexp_conv_lib.Sexp.t

val compare_expr_with : m f. ('m -> 'm -> Core_kernel__.Import.int) -> ('f -> 'f -> Core_kernel__.Import.int) -> ('m, 'f) expr_with -> ('m, 'f) expr_with -> Core_kernel__.Import.int

val hash_fold_expr_with : m f. (Ppx_hash_lib.Std.Hash.state -> 'm -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'f -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('m, 'f) expr_with -> Ppx_hash_lib.Std.Hash.state

type located_meta = {

loc : Middle.Location_span.t;

}

Untyped expressions, which have location_spans as meta-data

val map_located_meta : located_meta -> located_meta

val fold_located_meta : 'a -> located_meta -> 'a

val located_meta_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> located_meta

val sexp_of_located_meta : located_meta -> Ppx_sexp_conv_lib.Sexp.t

val compare_located_meta : located_meta -> located_meta -> int

val hash_fold_located_meta : Ppx_hash_lib.Std.Hash.state -> located_meta -> Ppx_hash_lib.Std.Hash.state

val hash_located_meta : located_meta -> Ppx_hash_lib.Std.Hash.hash_value

type untyped_expression = (located_meta, unit) expr_with

val map_untyped_expression : 'a -> 'a

val fold_untyped_expression : 'a -> 'b -> 'a

val untyped_expression_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> untyped_expression

val sexp_of_untyped_expression : untyped_expression -> Ppx_sexp_conv_lib.Sexp.t

val compare_untyped_expression : untyped_expression -> untyped_expression -> int

val hash_fold_untyped_expression : Ppx_hash_lib.Std.Hash.state -> untyped_expression -> Ppx_hash_lib.Std.Hash.state

val hash_untyped_expression : untyped_expression -> Ppx_hash_lib.Std.Hash.hash_value

type typed_expr_meta = {

loc : Middle.Location_span.t;

ad_level : Middle.UnsizedType.autodifftype;

type_ : Middle.UnsizedType.t;

}

Typed expressions also have meta-data after type checking: a location_span, as well as a type and an origin block (lub of the origin blocks of the identifiers in it)

val map_typed_expr_meta : typed_expr_meta -> typed_expr_meta

val fold_typed_expr_meta : 'a -> typed_expr_meta -> 'a

val typed_expr_meta_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> typed_expr_meta

val sexp_of_typed_expr_meta : typed_expr_meta -> Ppx_sexp_conv_lib.Sexp.t

val compare_typed_expr_meta : typed_expr_meta -> typed_expr_meta -> int

val hash_fold_typed_expr_meta : Ppx_hash_lib.Std.Hash.state -> typed_expr_meta -> Ppx_hash_lib.Std.Hash.state

val hash_typed_expr_meta : typed_expr_meta -> Ppx_hash_lib.Std.Hash.hash_value

type typed_expression = (typed_expr_meta, fun_kind) expr_with

val map_typed_expression : 'a -> 'a

val fold_typed_expression : 'a -> 'b -> 'a

val typed_expression_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> typed_expression

val sexp_of_typed_expression : typed_expression -> Ppx_sexp_conv_lib.Sexp.t

val compare_typed_expression : typed_expression -> typed_expression -> int

val hash_fold_typed_expression : Ppx_hash_lib.Std.Hash.state -> typed_expression -> Ppx_hash_lib.Std.Hash.state

val hash_typed_expression : typed_expression -> Ppx_hash_lib.Std.Hash.hash_value

val mk_untyped_expression : expr:((located_meta, 'a) expr_with, 'a) expression -> loc:Middle.Location_span.t -> (located_meta, 'a) expr_with

val mk_typed_expression : expr:((typed_expr_meta, 'a) expr_with, 'a) expression -> loc:Middle.Location_span.t -> type_:Middle.UnsizedType.t -> ad_level:Middle.UnsizedType.autodifftype -> (typed_expr_meta, 'a) expr_with

val expr_loc_lub : (typed_expr_meta, 'a) expr_with list -> Middle.Location_span.t

val expr_ad_lub : (typed_expr_meta, 'a) expr_with list -> Middle.UnsizedType.autodifftype

Least upper bound of expression autodiff types

type assignmentoperator =

| Assign

| ArrowAssign

| OperatorAssign of Middle.Operator.t

Assignment operators

val assignmentoperator_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> assignmentoperator

val sexp_of_assignmentoperator : assignmentoperator -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_assignmentoperator : Ppx_hash_lib.Std.Hash.state -> assignmentoperator -> Ppx_hash_lib.Std.Hash.state

val hash_assignmentoperator : assignmentoperator -> Ppx_hash_lib.Std.Hash.hash_value

val compare_assignmentoperator : assignmentoperator -> assignmentoperator -> int

type 'e truncation =

| NoTruncate

| TruncateUpFrom of 'e

| TruncateDownFrom of 'e

| TruncateBetween of 'e * 'e

Truncations

val map_truncation : ('a -> 'b) -> 'a truncation -> 'b truncation

val fold_truncation : ('a -> 'b -> 'a) -> 'a -> 'b truncation -> 'a

val truncation_of_sexp : e. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> Ppx_sexp_conv_lib.Sexp.t -> 'e truncation

val sexp_of_truncation : e. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> 'e truncation -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_truncation : e. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'e truncation -> Ppx_hash_lib.Std.Hash.state

val compare_truncation : e. ('e -> 'e -> int) -> 'e truncation -> 'e truncation -> int

type 'e printable =

| PString of string

| PExpr of 'e

Things that can be printed

val map_printable : ('a -> 'b) -> 'a printable -> 'b printable

val fold_printable : ('a -> 'b -> 'a) -> 'a -> 'b printable -> 'a

val printable_of_sexp : e. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> Ppx_sexp_conv_lib.Sexp.t -> 'e printable

val sexp_of_printable : e. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> 'e printable -> Ppx_sexp_conv_lib.Sexp.t

val compare_printable : e. ('e -> 'e -> Core_kernel__.Import.int) -> 'e printable -> 'e printable -> Core_kernel__.Import.int

val hash_fold_printable : e. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'e printable -> Ppx_hash_lib.Std.Hash.state

type ('l, 'e) lvalue =

| LVariable of identifier

| LIndexed of 'l * 'e index list

val map_lvalue : ('a -> 'b) -> ('c -> 'd) -> ('a, 'c) lvalue -> ('b, 'd) lvalue

val fold_lvalue : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> 'a -> ('b, 'c) lvalue -> 'a

val lvalue_of_sexp : l e. (Ppx_sexp_conv_lib.Sexp.t -> 'l) -> (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> Ppx_sexp_conv_lib.Sexp.t -> ('l, 'e) lvalue

val sexp_of_lvalue : l e. ('l -> Ppx_sexp_conv_lib.Sexp.t) -> ('e -> Ppx_sexp_conv_lib.Sexp.t) -> ('l, 'e) lvalue -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_lvalue : l e. (Ppx_hash_lib.Std.Hash.state -> 'l -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('l, 'e) lvalue -> Ppx_hash_lib.Std.Hash.state

val compare_lvalue : l e. ('l -> 'l -> Core_kernel__.Import.int) -> ('e -> 'e -> Core_kernel__.Import.int) -> ('l, 'e) lvalue -> ('l, 'e) lvalue -> Core_kernel__.Import.int

type ('e, 'm) lval_with = {

lval : (('e, 'm) lval_with, 'e) lvalue;

lmeta : 'm;

}

val map_lval_with : ('e -> 'a) -> ('m -> 'b) -> ('e, 'm) lval_with -> ('a, 'b) lval_with

val fold_lval_with : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> 'a -> ('b, 'c) lval_with -> 'a

val lval_with_of_sexp : e m. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> (Ppx_sexp_conv_lib.Sexp.t -> 'm) -> Ppx_sexp_conv_lib.Sexp.t -> ('e, 'm) lval_with

val sexp_of_lval_with : e m. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> ('m -> Ppx_sexp_conv_lib.Sexp.t) -> ('e, 'm) lval_with -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_lval_with : e m. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'm -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('e, 'm) lval_with -> Ppx_hash_lib.Std.Hash.state

val compare_lval_with : e m. ('e -> 'e -> Core_kernel__.Import.int) -> ('m -> 'm -> Core_kernel__.Import.int) -> ('e, 'm) lval_with -> ('e, 'm) lval_with -> Core_kernel__.Import.int

type untyped_lval = (untyped_expression, located_meta) lval_with

val map_untyped_lval : 'a -> 'a

val fold_untyped_lval : 'a -> 'b -> 'a

val untyped_lval_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> untyped_lval

val sexp_of_untyped_lval : untyped_lval -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_untyped_lval : Ppx_hash_lib.Std.Hash.state -> untyped_lval -> Ppx_hash_lib.Std.Hash.state

val hash_untyped_lval : untyped_lval -> Ppx_hash_lib.Std.Hash.hash_value

val compare_untyped_lval : untyped_lval -> untyped_lval -> int

type typed_lval = (typed_expression, typed_expr_meta) lval_with

val map_typed_lval : 'a -> 'a

val fold_typed_lval : 'a -> 'b -> 'a

val typed_lval_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> typed_lval

val sexp_of_typed_lval : typed_lval -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_typed_lval : Ppx_hash_lib.Std.Hash.state -> typed_lval -> Ppx_hash_lib.Std.Hash.state

val hash_typed_lval : typed_lval -> Ppx_hash_lib.Std.Hash.hash_value

val compare_typed_lval : typed_lval -> typed_lval -> int

type ('e, 's, 'l, 'f) statement =

| Assignment of {

assign_lhs : 'l;

assign_op : assignmentoperator;

assign_rhs : 'e;

}

| NRFunApp of 'f * identifier * 'e list

| TargetPE of 'e

| IncrementLogProb of 'e

| Tilde of {

arg : 'e;

distribution : identifier;

args : 'e list;

truncation : 'e truncation;

}

| Break

| Continue

| Return of 'e

| ReturnVoid

| Print of 'e printable list

| Reject of 'e printable list

| Skip

| IfThenElse of 'e * 's * 's option

| While of 'e * 's

| For of {

loop_variable : identifier;

lower_bound : 'e;

upper_bound : 'e;

loop_body : 's;

}

| ForEach of identifier * 'e * 's

| Profile of string * 's list

| Block of 's list

| VarDecl of {

decl_type : 'e Middle.Type.t;

transformation : 'e Middle.Transformation.t;

identifier : identifier;

initial_value : 'e option;

is_global : bool;

}

| FunDef of {

returntype : Middle.UnsizedType.returntype;

funname : identifier;

arguments : (Middle.UnsizedType.autodifftype * Middle.UnsizedType.t * identifier) list;

body : 's;

}

Statement shapes, where we substitute untyped_expression and untyped_statement for 'e and 's respectively to get untyped_statement and typed_expression and typed_statement to get typed_statement

val map_statement : ('a -> 'b) -> ('c -> 'd) -> ('e -> 'f) -> ('g -> 'h) -> ('a, 'c, 'e, 'g) statement -> ('b, 'd, 'f, 'h) statement

val fold_statement : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> ('a -> 'd -> 'a) -> ('a -> 'e -> 'a) -> 'a -> ('b, 'c, 'd, 'e) statement -> 'a

val statement_of_sexp : e s l f. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> (Ppx_sexp_conv_lib.Sexp.t -> 's) -> (Ppx_sexp_conv_lib.Sexp.t -> 'l) -> (Ppx_sexp_conv_lib.Sexp.t -> 'f) -> Ppx_sexp_conv_lib.Sexp.t -> ('e, 's, 'l, 'f) statement

val sexp_of_statement : e s l f. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> ('s -> Ppx_sexp_conv_lib.Sexp.t) -> ('l -> Ppx_sexp_conv_lib.Sexp.t) -> ('f -> Ppx_sexp_conv_lib.Sexp.t) -> ('e, 's, 'l, 'f) statement -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_statement : e s l f. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 's -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'l -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'f -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('e, 's, 'l, 'f) statement -> Ppx_hash_lib.Std.Hash.state

val compare_statement : e s l f. ('e -> 'e -> Core_kernel__.Import.int) -> ('s -> 's -> Core_kernel__.Import.int) -> ('l -> 'l -> Core_kernel__.Import.int) -> ('f -> 'f -> Core_kernel__.Import.int) -> ('e, 's, 'l, 'f) statement -> ('e, 's, 'l, 'f) statement -> Core_kernel__.Import.int

type statement_returntype =

| NoReturnType

| Incomplete of Middle.UnsizedType.returntype

| Complete of Middle.UnsizedType.returntype

| AnyReturnType

Statement return types which we will decorate statements with during type checking: the purpose is to check that function bodies have the correct return type in every possible execution branch. NoReturnType corresponds to not having a return statement in it. Incomplete rt corresponds to having some return statement(s) of type rt in it, but not one in every branch Complete rt corresponds to having a return statement of type rt in every branch AnyReturnType corresponds to statements which have an error in every branch

val statement_returntype_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> statement_returntype

val sexp_of_statement_returntype : statement_returntype -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_statement_returntype : Ppx_hash_lib.Std.Hash.state -> statement_returntype -> Ppx_hash_lib.Std.Hash.state

val hash_statement_returntype : statement_returntype -> Ppx_hash_lib.Std.Hash.hash_value

val compare_statement_returntype : statement_returntype -> statement_returntype -> int

type ('e, 'm, 'l, 'f) statement_with = {

stmt : ('e, ('e, 'm, 'l, 'f) statement_with, 'l, 'f) statement;

smeta : 'm;

}

val map_statement_with : ('e -> 'a) -> ('m -> 'b) -> ('l -> 'c) -> ('f -> 'd) -> ('e, 'm, 'l, 'f) statement_with -> ('a, 'b, 'c, 'd) statement_with

val fold_statement_with : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> ('a -> 'd -> 'a) -> ('a -> 'e -> 'a) -> 'a -> ('b, 'c, 'd, 'e) statement_with -> 'a

val statement_with_of_sexp : e m l f. (Ppx_sexp_conv_lib.Sexp.t -> 'e) -> (Ppx_sexp_conv_lib.Sexp.t -> 'm) -> (Ppx_sexp_conv_lib.Sexp.t -> 'l) -> (Ppx_sexp_conv_lib.Sexp.t -> 'f) -> Ppx_sexp_conv_lib.Sexp.t -> ('e, 'm, 'l, 'f) statement_with

val sexp_of_statement_with : e m l f. ('e -> Ppx_sexp_conv_lib.Sexp.t) -> ('m -> Ppx_sexp_conv_lib.Sexp.t) -> ('l -> Ppx_sexp_conv_lib.Sexp.t) -> ('f -> Ppx_sexp_conv_lib.Sexp.t) -> ('e, 'm, 'l, 'f) statement_with -> Ppx_sexp_conv_lib.Sexp.t

val compare_statement_with : e m l f. ('e -> 'e -> Core_kernel__.Import.int) -> ('m -> 'm -> Core_kernel__.Import.int) -> ('l -> 'l -> Core_kernel__.Import.int) -> ('f -> 'f -> Core_kernel__.Import.int) -> ('e, 'm, 'l, 'f) statement_with -> ('e, 'm, 'l, 'f) statement_with -> Core_kernel__.Import.int

val hash_fold_statement_with : e m l f. (Ppx_hash_lib.Std.Hash.state -> 'e -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'm -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'l -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'f -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('e, 'm, 'l, 'f) statement_with -> Ppx_hash_lib.Std.Hash.state

type untyped_statement = (untyped_expression, located_meta, untyped_lval, unit) statement_with

Untyped statements, which have location_spans as meta-data

val map_untyped_statement : 'a -> 'a

val untyped_statement_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> untyped_statement

val sexp_of_untyped_statement : untyped_statement -> Ppx_sexp_conv_lib.Sexp.t

val compare_untyped_statement : untyped_statement -> untyped_statement -> int

val hash_fold_untyped_statement : Ppx_hash_lib.Std.Hash.state -> untyped_statement -> Ppx_hash_lib.Std.Hash.state

val hash_untyped_statement : untyped_statement -> Ppx_hash_lib.Std.Hash.hash_value

val mk_untyped_statement : stmt:(untyped_expression, (untyped_expression, located_meta, untyped_lval, unit) statement_with, untyped_lval, unit) statement -> loc:Middle.Location_span.t -> untyped_statement

type stmt_typed_located_meta = {

loc : Middle.Location_span.t;

return_type : statement_returntype;

}

val map_stmt_typed_located_meta : stmt_typed_located_meta -> stmt_typed_located_meta

val stmt_typed_located_meta_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> stmt_typed_located_meta

val sexp_of_stmt_typed_located_meta : stmt_typed_located_meta -> Ppx_sexp_conv_lib.Sexp.t

val compare_stmt_typed_located_meta : stmt_typed_located_meta -> stmt_typed_located_meta -> int

val hash_fold_stmt_typed_located_meta : Ppx_hash_lib.Std.Hash.state -> stmt_typed_located_meta -> Ppx_hash_lib.Std.Hash.state

val hash_stmt_typed_located_meta : stmt_typed_located_meta -> Ppx_hash_lib.Std.Hash.hash_value

type typed_statement = (typed_expression, stmt_typed_located_meta, typed_lval, fun_kind) statement_with

Typed statements also have meta-data after type checking: a location_span, as well as a statement returntype to check that function bodies have the right return type

val map_typed_statement : 'a -> 'a

val typed_statement_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> typed_statement

val sexp_of_typed_statement : typed_statement -> Ppx_sexp_conv_lib.Sexp.t

val compare_typed_statement : typed_statement -> typed_statement -> int

val hash_fold_typed_statement : Ppx_hash_lib.Std.Hash.state -> typed_statement -> Ppx_hash_lib.Std.Hash.state

val hash_typed_statement : typed_statement -> Ppx_hash_lib.Std.Hash.hash_value

val mk_typed_statement : stmt:('a, ('a, stmt_typed_located_meta, 'b, 'c) statement_with, 'b, 'c) statement -> loc:Middle.Location_span.t -> return_type:statement_returntype -> ('a, stmt_typed_located_meta, 'b, 'c) statement_with

type 's block = {

stmts : 's list;

xloc : Middle.Location_span.t;

}

Program shapes, where we obtain types of programs if we substitute typed or untyped statements for 's

and comment_type =

| LineComment of string * Middle.Location_span.t

| Include of string * Middle.Location_span.t

| BlockComment of string list * Middle.Location_span.t

| Separator of Middle.Location.t
Separator records the location of items like commas, operators, and keywords which don't have location information stored in the AST but are useful for placing comments in pretty printing

and 's program = {

functionblock : 's block option;

datablock : 's block option;

transformeddatablock : 's block option;

parametersblock : 's block option;

transformedparametersblock : 's block option;

modelblock : 's block option;

generatedquantitiesblock : 's block option;

comments : comment_type list;

}

val map_block : ('s -> 'a) -> 's block -> 'a block

val map_comment_type : comment_type -> comment_type

val map_program : ('s -> 'a) -> 's program -> 'a program

val fold_block : ('a -> 'b -> 'a) -> 'a -> 'b block -> 'a

val fold_comment_type : 'a -> comment_type -> 'a

val fold_program : ('a -> 'b -> 'a) -> 'a -> 'b program -> 'a

val block_of_sexp : s. (Ppx_sexp_conv_lib.Sexp.t -> 's) -> Ppx_sexp_conv_lib.Sexp.t -> 's block

val comment_type_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> comment_type

val program_of_sexp : s. (Ppx_sexp_conv_lib.Sexp.t -> 's) -> Ppx_sexp_conv_lib.Sexp.t -> 's program

val sexp_of_block : s. ('s -> Ppx_sexp_conv_lib.Sexp.t) -> 's block -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_comment_type : comment_type -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_program : s. ('s -> Ppx_sexp_conv_lib.Sexp.t) -> 's program -> Ppx_sexp_conv_lib.Sexp.t

val hash_fold_block : s. (Ppx_hash_lib.Std.Hash.state -> 's -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 's block -> Ppx_hash_lib.Std.Hash.state

val hash_fold_comment_type : Ppx_hash_lib.Std.Hash.state -> comment_type -> Ppx_hash_lib.Std.Hash.state

val hash_fold_program : s. (Ppx_hash_lib.Std.Hash.state -> 's -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 's program -> Ppx_hash_lib.Std.Hash.state

val hash_comment_type : comment_type -> Ppx_hash_lib.Std.Hash.hash_value

val compare_block : s. ('s -> 's -> Core_kernel__.Import.int) -> 's block -> 's block -> Core_kernel__.Import.int

val compare_comment_type : comment_type -> comment_type -> int

val compare_program : s. ('s -> 's -> Core_kernel__.Import.int) -> 's program -> 's program -> Core_kernel__.Import.int

val get_stmts : 'a block option -> 'a list

type untyped_program = untyped_statement program

Untyped programs (before type checking)

val map_untyped_program : 'a -> 'a

val untyped_program_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> untyped_program

val sexp_of_untyped_program : untyped_program -> Ppx_sexp_conv_lib.Sexp.t

val compare_untyped_program : untyped_program -> untyped_program -> int

type typed_program = typed_statement program

Typed programs (after type checking)

val map_typed_program : 'a -> 'a

val typed_program_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> typed_program

val sexp_of_typed_program : typed_program -> Ppx_sexp_conv_lib.Sexp.t

val compare_typed_program : typed_program -> typed_program -> int

val untyped_expression_of_typed_expression : typed_expression -> untyped_expression

Forgetful function from typed to untyped expressions

val untyped_lvalue_of_typed_lvalue : typed_lval -> untyped_lval

val untyped_statement_of_typed_statement : (typed_expression, stmt_typed_located_meta, typed_lval, 'a) statement_with -> (untyped_expression, located_meta, untyped_lval, unit) statement_with

Forgetful function from typed to untyped statements

val untyped_program_of_typed_program : typed_program -> untyped_program

Forgetful function from typed to untyped programs

val expr_of_lvalue : (('a, 'b) expr_with, 'a) lval_with -> ('a, 'b) expr_with

val lvalue_of_expr : ('a, 'b) expr_with -> (('a, 'b) expr_with, 'a) lval_with

val id_of_lvalue : ('a, 'b) lval_with -> identifier

val get_loc_expr : untyped_expression -> Middle.Location.t

val get_loc_dt : untyped_expression Middle.Type.t -> Middle.Location.t option

val get_loc_tf : untyped_expression Middle.Transformation.t -> Middle.Location.t option

val get_first_loc : untyped_statement -> Middle.Location.t