Module Frontend.Errors

type syntax_error =

| Lexing of Middle.Location.t

| UnexpectedEOF of Middle.Location.t

| Include of string * Middle.Location.t

| Parsing of string * Middle.Location_span.t

Our type of syntax error information

exception SyntaxError of syntax_error

Exception for Syntax Errors

exception SemanticError of Semantic_error.t

Exception SemanticError (loc, msg) indicates a semantic error with message msg, occurring at location loc.

type t =

| FileNotFound of string

| Syntax_error of syntax_error

| Semantic_error of Semantic_error.t

val pp : ?⁠printed_filename:string -> t Fmt.t

val to_string : t -> string

val pp_syntax_error : ?⁠printed_filename:string -> Stdlib.Format.formatter -> syntax_error -> unit

A syntax error message used when handling a SyntaxError

val pp_semantic_error : ?⁠printed_filename:string -> Stdlib.Format.formatter -> Semantic_error.t -> unit

A semantic error message used when handling a SemanticError