Module Middle

Modules which define the 'middle intermediate representation' (MIR) of Stan

module Location : sig ... end

Storing locations in the original source

module Location_span : sig ... end

Delimited locations in source code

module Operator : sig ... end

Utilities for Stan's built in operators

module Fun_kind : sig ... end

Types for function kinds, e.g. StanLib or UserDefined, and function suffix types, e.g. foo_ldfp, bar_lp

module Flag_vars : sig ... end

module Internal_fun : sig ... end

Language functions defined internally by the compiler

module Index : sig ... end

Types of indexing operations

module Expr : sig ... end

MIR types and modules corresponding to the expressions of the language

module UnsizedType : sig ... end

Types which have dimensionalities but not sizes, e.g. array[,,]

module SizedType : sig ... end

Types which have a concrete size associated, e.g. vector[n]

module Type : sig ... end

A type which unifies SizedTypes and UnsizedTypes for declarations

module Stmt : sig ... end

MIR types and modules corresponding to the statements of the language

module Program : sig ... end

Defines the core of the MIR

module Stan_math_signatures : sig ... end

module Utils : sig ... end

Utilities, primarily surrounding distribution names and suffixes

module Transformation : sig ... end

Transformations (constraints) for global variable declarations