Module Middle.SizedType

Types which have a concrete size associated, e.g. vector[n]

type 'a t =

| SInt

| SReal

| SComplex

| SVector of Common.Helpers.mem_pattern * 'a

| SRowVector of Common.Helpers.mem_pattern * 'a

| SMatrix of Common.Helpers.mem_pattern * 'a * 'a

| SArray of 'a t * 'a

val map : ('a -> 'b) -> 'a t -> 'b t

val fold : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

val t_of_sexp : a. (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a t

val sexp_of_t : a. ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val compare : a. ('a -> 'a -> int) -> 'a t -> 'a t -> int

val hash_fold_t : a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a t -> Ppx_hash_lib.Std.Hash.state

val pp : 'a Fmt.t -> Stdlib.Format.formatter -> 'a t -> unit

val collect_exprs : 'a t -> 'a list

val to_unsized : 'a t -> UnsizedType.t

val associate : ?⁠init:Expr.Labelled.t Common.Label.Int_label.Map.t -> Expr.Labelled.t t -> Expr.Labelled.t Common.Label.Int_label.Map.t

val inner_type : 'a t -> 'a t

val contains_complex : 'a t -> bool

val dims_of : 'a t -> 'a list

val get_dims_io : Middle__Expr.Typed.t t -> Middle__Expr.Typed.t list

Get the dimensions with respect to sizes needed for IO. Note: The main difference from get_dims is complex, where this function treats the complex type as a dual number.

val get_dims : 'a t -> 'a list

val is_recursive_container : 'a t -> bool

* Check whether a SizedType holds indexable SizedTypes.

val get_array_dims : 'a t -> 'a t * 'a list

Return a type's array dimensions and the type inside the (possibly nested) array

val num_elems_expr : Middle__Expr.Typed.t t -> Middle__Expr.Typed.t

val contains_eigen_type : 'a t -> bool

* Return true if SizedType contains an Eigen type

val contains_soa : 'a t -> bool

* Return true if SizedType contains a type tagged SoA

val get_mem_pattern : 'a t -> Common.Helpers.mem_pattern

* Return the mem_pattern of the SizedType

val demote_sizedtype_mem : 'a t -> 'a t

val promote_sizedtype_mem : 'a t -> 'a t

val modify_sizedtype_mem : Common.Helpers.mem_pattern -> 'a t -> 'a t